+++ src/main/java/com/umc/networkingService/domain/member/service/MemberService.java
@@ -1,6 +1,5 @@
 package com.umc.networkingService.domain.member.service;
 
-import com.umc.networkingService.domain.member.dto.request.MemberSignUpRequest;
 import com.umc.networkingService.domain.member.dto.request.MemberUpdateMyProfileRequest;
 import com.umc.networkingService.domain.member.dto.request.MemberUpdateProfileRequest;
 import com.umc.networkingService.domain.member.dto.response.*;
@@ -11,10 +10,6 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.UUID;
 
 public interface MemberService extends EntityLoader<Member, UUID> {
-    MemberIdResponse signUp(Member member, MemberSignUpRequest request);
-    MemberGenerateNewAccessTokenResponse generateNewAccessToken(String refreshToken, Member member);
-    MemberIdResponse logout(Member member);
-    MemberIdResponse withdrawal(Member member);
     MemberIdResponse updateMyProfile(Member member, MultipartFile profileImage, MemberUpdateMyProfileRequest request);
     MemberIdResponse updateProfile(Member member, UUID memberId, MemberUpdateProfileRequest request);
     MemberInquiryProfileResponse inquiryProfile(Member member, UUID memberId);
